[Feat/#91] 토론장 네트워크 연결 시작

+++ JYP-iOS/JYP-iOS/Sources/Model/DTO/CreatePikmiRequest.swift
@@ -10,6 +10,6 @@ import Foundation
 
 struct CreatePikmiRequest: Codable {
     let name, address, category: String
-    let longitude, latitude: Int
+    let longitude, latitude: Double
     let link: String
 }
